+++ jaxrs-base-server/src/test/java/io/telicent/smart/cache/server/jaxrs/model/TestProblem.java
@@ -17,23 +17,36 @@ package io.telicent.smart.cache.server.jaxrs.model;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.ws.rs.core.HttpHeaders;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import org.testng.Assert;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
+
+import java.util.List;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class TestProblem {
 
     private final ObjectMapper json = new ObjectMapper();
 
     @Test
-    public void problem_test_01() throws JsonProcessingException {
-        Problem problem = new Problem("Test", "Title", 400, "You were being bad", "BadRequest#1234");
+    public void givenProblemDetails_whenConstructingProblem_thenValuesAsExpected_andRoundTripsViaJson() throws
+            JsonProcessingException {
+        // Given and When
+        Problem problem = createTestProblem();
+
+        // Then
         Assert.assertTrue(problem.equals(problem));
         Assert.assertFalse(problem.equals(null));
         Assert.assertFalse(problem.equals(new Object()));
 
+        // And
         String json = this.json.writeValueAsString(problem);
         Problem retrieved = this.json.readValue(json, Problem.class);
-
         Assert.assertEquals(retrieved, problem);
         Assert.assertTrue(retrieved.equals(problem));
         Assert.assertEquals(retrieved.hashCode(), problem.hashCode());
@@ -41,13 +54,16 @@ public class TestProblem {
     }
 
     @Test
-    public void problem_test_02() {
-        Problem a = new Problem("Test", "Title", 400, "You were being bad", "BadRequest#1234");
-        Problem b = new Problem("Test", "Title", 400, "You were being bad", "BadRequest#1234");
+    public void givenTwoIdenticalProblems_whenComparingEquality_thenCorrect_andMutatingOneChangesEqualityResult() {
+        // Given
+        Problem a = createTestProblem();
+        Problem b = createTestProblem();
 
+        // When and Then
         Assert.assertEquals(a, b);
         Assert.assertTrue(a.equals(b));
 
+        // And
         // Mutate each field checking equality no longer holds
         b.setDetail(null);
         verifyNotEquals(a, b);
@@ -64,5 +80,69 @@ public class TestProblem {
     private static void verifyNotEquals(Problem a, Problem b) {
         Assert.assertNotEquals(a, b);
         Assert.assertFalse(a.equals(b));
+    }
+
+    @Test
+    public void givenProblem_whenConvertingToResponse_thenDefaultContentTypeSet() {
+        // Given
+        Problem problem = createTestProblem();
+
+        // When
+        Response response = problem.toResponse();
+
+        // Then
+        Assert.assertEquals(response.getMediaType(), Problem.APPLICATION_PROBLEM_JSON);
+    }
+
+    @Test
+    public void givenProblem_whenConvertingToResponseWithWildcardAccept_thenDefaultContentTypeSet() {
+        // Given
+        Problem problem = createTestProblem();
+        HttpHeaders headers = mock(HttpHeaders.class);
+        when(headers.getAcceptableMediaTypes()).thenReturn(List.of(MediaType.WILDCARD_TYPE));
+
+        // When
+        Response response = problem.toResponse(headers);
+
+        // Then
+        Assert.assertEquals(response.getMediaType(), Problem.APPLICATION_PROBLEM_JSON);
+    }
+
+    @DataProvider(name = "mediaTypes")
+    private Object[][] mediaTypes() {
+        return new Object[][] {
+                {
+                        List.of(MediaType.APPLICATION_JSON, Problem.MEDIA_TYPE), MediaType.APPLICATION_JSON
+                }, {
+                        List.of(MediaType.TEXT_PLAIN), MediaType.TEXT_PLAIN
+                }, {
+                        List.of(Problem.MEDIA_TYPE), Problem.MEDIA_TYPE
+                }, {
+                        List.of("application/custom"), null
+                }
+        };
+    }
+
+    @Test(dataProvider = "mediaTypes")
+    public void givenProblem_whenConvertingToResponseWithAccept_thenExpectedContentTypeSet(List<String> accept,
+                                                                                           String expected) {
+        // Given
+        Problem problem = createTestProblem();
+        HttpHeaders headers = mock(HttpHeaders.class);
+        when(headers.getAcceptableMediaTypes()).thenReturn(accept.stream().map(MediaType::valueOf).toList());
+
+        // When
+        Response response = problem.toResponse(headers);
+
+        // Then
+        if (expected != null) {
+            Assert.assertEquals(response.getMediaType(), MediaType.valueOf(expected));
+        } else {
+            Assert.assertNull(response.getMediaType());
+        }
+    }
+
+    private static Problem createTestProblem() {
+        return new Problem("Test", "Title", 400, "You were being bad", "BadRequest#1234");
     }
 }